♻️ Refactor: strongly-type Gcal and SocketErrors
