+++ packages/backend/src/common/constants/error.constants.ts
@@ -1,20 +1,6 @@
 import { ErrorConstant } from "@core/errors/errors.base";
 import { Status } from "@core/errors/status.codes";
 
-export const SocketError: ErrorConstant = {
-  InvalidSocketId: {
-    description: "Invalid socket id",
-    status: Status.BAD_REQUEST,
-    isOperational: true,
-  },
-  ServerNotReady: {
-    description:
-      "WebSocket server not ready (Did you forget to initialize it?)",
-    status: Status.INTERNAL_SERVER,
-    isOperational: false,
-  },
-};
-
 export const UserError: ErrorConstant = {
   InvalidValue: {
     description: "User has an invalid value",
